requirements.txt
Django
whitenoise
azure-search-documents --pre
python-dotenv
nltk
beautifulsoup4